FIX #17915:Introduce Icecast for macOS
Introduces Icecast as a new entry for the streaming option.
There are 3 new entries in Convert & Stream menu
	- Mount Point: Here user needs to enter mountpoint.
	- Username	 : Here user needs to enter his username.
	- Password	 : Here user needs to enter his password.
In other versions of VLCs, there are no separate entries for username and password
instead
In other versions of VLCs which support icecast,there is only a  single entry
"Login:pass" for inputting username and password for icecast which is really hard for
user to input,to prevent that there is separate entry to input username and password
this will  make more user-friendly and enriches the User Experience.
resolves: #17915.